fix: three correctness issues — dead null check, Dispose idempotency, idxDiv overflow
StreamDecoder.Read(Span<float>): remove dead null guard on the buffer
parameter.  Span<float> is a struct; the compiler rejects null assignment,
so the check could never fire and confused readers into thinking the method
might accept a null-like value.

VorbisReader.Dispose: add _disposed flag so double-Dispose is a strict
no-op.  Previously the readonly _decoders / _containerReader fields meant
neither null check in Dispose could ever be false, so a second call would
re-invoke _containerReader.Dispose() and re-iterate the cleared decoder list.

Codebook.InitLookupTable: change idxDiv from int to long.  The variable
accumulates idxDiv *= lookupValueCount across Dimensions iterations; when
Dimensions > ~9 and lookupValueCount > 1 the product silently wraps,
producing a wrong moff index and corrupting the lookup table for that
codebook entry.

+++ NVorbis/Codebook.cs
@@ -249,10 +249,10 @@ namespace NVorbis
                 for (var idx = 0; idx < Entries; idx++)
                 {
                     var last = 0.0;
-                    var idxDiv = 1;
+                    var idxDiv = 1L;
                     for (var i = 0; i < Dimensions; i++)
                     {
-                        var moff = (idx / idxDiv) % lookupValueCount;
+                        var moff = (int)((idx / idxDiv) % lookupValueCount);
                         var value = (float)multiplicands[moff] * deltaValue + minValue + last;
                         lookupTable[idx * Dimensions + i] = (float)value;
